Etckeeper need to be installed in target rootfs
Since etckeeper files on target is used on rootfs creation time, etckeeper need
to be installed in target rootfs. Therefore, etckeeper is included in the
IMAGE_INSTALL variable.

Might consider in the future removing this and making the deployed /etc git
repository to be created without the need for etckeeper on target.